Jamendo: decrypter fixed ComboBrowseFile: überarbeitet damit man ihn in einen TreeTable einpflegen kann LinkGrabberV2: ersta alpha vom neuen linkgrabber
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@4982 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/FilePackage.java
@@ -321,7 +321,7 @@ public class FilePackage extends Property implements Serializable {
      * 
      * @return
      */
-    public int getTotalEstimatedPackageSize() {
+    public long getTotalEstimatedPackageSize() {
         if (System.currentTimeMillis() - updateTime > UPDATE_INTERVAL) {
             updateCollectives();
         }
